# application/app/Modules/Identity/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация API - модуля Identity",
        "version": "1.0.0"
    },
    "paths": {
        "/api/identity/auth/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Текущий пользователь со всеми разрешениями",
                "operationId": "982eca91964db2aa2831a53865e8a537",
                "responses": {
                    "200": {
                        "description": "Данные пользователя с полным списком разрешений"
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/identity/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Авторизация",
                "operationId": "82ab0a88cf68acf2652c2be1caf087e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная авторизация"
                    },
                    "422": {
                        "description": "Ошибка валидации / неверные данные"
                    }
                }
            }
        },
        "/api/identity/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Выход",
                "operationId": "f6302e8011f614198d60d63dc3ec0219",
                "responses": {
                    "200": {
                        "description": "Успешный выход"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer Token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}